Remove AbstractTestDecimalFunctions
We need to test operators that return decimal not only in dedicated `DECIMAL` test classes.
