--- presto-main/src/test/java/com/facebook/presto/type/AbstractTestDecimalFunctions.java
@@ -1,55 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.facebook.presto.type;
-
-import com.facebook.presto.operator.scalar.AbstractTestFunctions;
-import com.facebook.presto.spi.type.DecimalParseResult;
-import com.facebook.presto.spi.type.Decimals;
-import com.facebook.presto.spi.type.SqlDecimal;
-import io.airlift.slice.Slice;
-
-import java.math.BigInteger;
-
-import static com.facebook.presto.spi.type.DecimalType.createDecimalType;
-import static com.facebook.presto.spi.type.Decimals.MAX_PRECISION;
-
-public abstract class AbstractTestDecimalFunctions
-        extends AbstractTestFunctions
-{
-    protected void assertDecimalFunction(String statement, SqlDecimal expectedResult)
-    {
-        assertFunction(statement,
-                createDecimalType(expectedResult.getPrecision(), expectedResult.getScale()),
-                expectedResult);
-    }
-
-    protected SqlDecimal decimal(String decimalString)
-    {
-        DecimalParseResult parseResult = Decimals.parseIncludeLeadingZerosInPrecision(decimalString);
-        BigInteger unscaledValue;
-        if (parseResult.getType().isShort()) {
-            unscaledValue = BigInteger.valueOf((Long) parseResult.getObject());
-        }
-        else {
-            unscaledValue = Decimals.decodeUnscaledValue((Slice) parseResult.getObject());
-        }
-        return new SqlDecimal(unscaledValue, parseResult.getType().getPrecision(), parseResult.getType().getScale());
-    }
-
-    protected SqlDecimal maxPrecisionDecimal(long value)
-    {
-        final String maxPrecisionFormat = "%0" + (MAX_PRECISION + (value < 0 ? 1 : 0)) + "d";
-        return decimal(String.format(maxPrecisionFormat, value));
-    }
-}
