+++ presto-main/src/test/java/com/facebook/presto/type/TestDecimalParametricType.java
@@ -14,10 +14,11 @@
 
 package com.facebook.presto.type;
 
+import com.facebook.presto.operator.scalar.AbstractTestFunctions;
 import org.testng.annotations.Test;
 
 public class TestDecimalParametricType
-        extends AbstractTestDecimalFunctions
+        extends AbstractTestFunctions
 {
     @Test
     public void decimalIsCreatedWithPrecisionAndScaleDefined()
