+++ presto-main/src/test/java/com/facebook/presto/type/TestDecimalToDecimalCasts.java
@@ -13,10 +13,11 @@
  */
 package com.facebook.presto.type;
 
+import com.facebook.presto.operator.scalar.AbstractTestFunctions;
 import org.testng.annotations.Test;
 
 public class TestDecimalToDecimalCasts
-        extends AbstractTestDecimalFunctions
+        extends AbstractTestFunctions
 {
     @Test
     public void testShortDecimalToShortDecimalCasts()
